1. Changed Range class: switched to primitives 2. Changed REST API methods: added ResponseEntity, added exception 	handling 3. Changed RangeServiceTest

+++ src/main/java/com/wsi/range/controller/RangeController.java
@@ -1,15 +1,21 @@
 package com.wsi.range.controller;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+import com.wsi.range.exeption.RangeException;
 import com.wsi.range.model.Range;
+import com.wsi.range.model.Response;
 import com.wsi.range.service.RangeService;
 
 @RestController
@@ -26,26 +32,38 @@ public class RangeController
 		return "Just testing!";
 	}
  
-	@GetMapping({ "/ranges" })
-	public List<Range> listRanges()
+	@RequestMapping(value = "/ranges", method=RequestMethod.GET)
+	public ResponseEntity<List<Range>> listRanges()
 	{
-		log.debug("Listing ALL ranges...");
-		List<Range> ranges = service.listAll();
-		return ranges;
+		log.debug("Listing ALL ranges...");		
+		return new ResponseEntity<List<Range>>(service.listAll(), HttpStatus.OK);
 	}
 
-	@RequestMapping({ "/ranges/{id}" })
-	public Range getRange(@PathVariable String id) throws Exception
+	@RequestMapping(value = "/ranges/{id}", method=RequestMethod.GET)
+	public ResponseEntity<Range> getRangeById(@PathVariable("id") String id) throws RangeException
 	{
 		log.debug("Getting {} range...", id);
-		return service.getOne(UUID.fromString(id)).orElseThrow(() -> new java.lang.Exception());
+		Optional<Range> range = service.getOne(UUID.fromString(id));
+	
+		if (range.isPresent())
+			return new ResponseEntity<Range>(range.get(), HttpStatus.OK);
+		else
+			throw new RangeException("Range doesn't exist.");
 	}
 
-	@RequestMapping({ "/ranges/delete/{id}" })
-	public void delete(@PathVariable String id)
+	@RequestMapping(value = "/ranges/delete/{id}", method = RequestMethod.DELETE)
+	public ResponseEntity<Response> deleteRangeById(@PathVariable("id") String id) throws RangeException
 	{	
 		log.debug("Deleting {} range...", id);
-		service.delete(UUID.fromString(id));
+		Optional<Range> range = service.getOne(UUID.fromString(id));
+		
+		if (range.isPresent())
+		{
+			service.delete(UUID.fromString(id));
+			return new ResponseEntity<Response>(new Response(HttpStatus.OK.value(), "Range has been deleted."), HttpStatus.OK);
+		}
+		else
+			throw new RangeException("Range to delete doesn't exist.");
 	}
 	
 	@RequestMapping({ "/ranges/deleteall" })
@@ -55,19 +73,19 @@ public class RangeController
 		service.deleteAll();
 	}
 	
-	@RequestMapping({ "/ranges/add/{left}/{right}" })
-	public Range add(@PathVariable Integer left, @PathVariable Integer right)
+	@RequestMapping(value = "/ranges/add/{left}/{right}", method = RequestMethod.POST)
+	public ResponseEntity<Range> saveRange(@PathVariable("left") int left, @PathVariable("right") int right) throws RangeException
 	{	
-		Range range = new Range(UUID.randomUUID(), left, right);
-		log.debug("Adding range: {} ...", range.toString());
-		service.saveOrUpdate(range);
-		return range;
+		if(left >= right)
+			throw new RangeException("Invalid range.");
+		
+		return new ResponseEntity<Range>(service.saveOrUpdate(new Range(UUID.randomUUID(), left, right)), HttpStatus.OK);
 	}
 	
 	@GetMapping({ "/ranges/reduced" })
-	public List<Range> listReducedRanges()
+	public ResponseEntity<List<Range>> listReducedRanges()
 	{
 		log.debug("Listing REDUCED ranges..."); 
-		return Range.processRanges(service.listAll());
+		return new ResponseEntity<List<Range>>(Range.processRanges(service.listAll()), HttpStatus.OK);
 	}
 }
